% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_catch.R
\name{fish_catch}
\alias{fish_catch}
\title{Fish Catch Function}
\usage{
fish_catch(fish)
}
\arguments{
\item{fish, }{a vector of fish species (names)}
}
\value{
list with the following items
\describe{
\item{most_common}{Name of the most frequently caught species}
\item{rarest}{Name of the least frequently caught species}
\item{total}{Total number of fish caught}
}
}
\description{
This function takes a vector of fish names and returns the most common fish, rarest fish, and total number of fish caught.
}
\author{
Sydney Rilum
}
